Avoid unused symbols on non-Linux
