+++ common/libnetwork/slirp4netns/const.go
@@ -3,14 +3,7 @@ package slirp4netns
 import "net"
 
 const (
-	ipv6ConfDefaultAcceptDadSysctl = "/proc/sys/net/ipv6/conf/default/accept_dad"
-	BinaryName                     = "slirp4netns"
-
-	// defaultMTU the default MTU override.
-	defaultMTU = 65520
-
-	// default slirp4ns subnet.
-	defaultSubnet = "10.0.2.0/24"
+	BinaryName = "slirp4netns"
 )
 
 // SetupResult return type from Setup().
