+++ common/pkg/cgroups/cgroups_unsupported.go
@@ -2,10 +2,6 @@
 
 package cgroups
 
-import (
-	"os"
-)
-
 // IsCgroup2UnifiedMode returns whether we are running in cgroup 2 cgroup2 mode.
 func IsCgroup2UnifiedMode() (bool, error) {
 	return false, nil
@@ -15,8 +11,4 @@ func IsCgroup2UnifiedMode() (bool, error) {
 // current cgroup.
 func UserOwnsCurrentSystemdCgroup() (bool, error) {
 	return false, nil
-}
-
-func rmDirRecursively(path string) error {
-	return os.RemoveAll(path)
 }
