+++ common/pkg/config/default_bsd.go
@@ -28,7 +28,3 @@ func getLibpodTmpDir() string {
 func getDefaultMachineVolumes() []string {
 	return []string{"$HOME:$HOME"}
 }
-
-func getDefaultComposeProviders() []string {
-	return defaultUnixComposeProviders
-}
