+++ common/pkg/config/default_linux.go
@@ -29,7 +29,3 @@ func getLibpodTmpDir() string {
 func getDefaultMachineVolumes() []string {
 	return []string{"$HOME:$HOME"}
 }
-
-func getDefaultComposeProviders() []string {
-	return defaultUnixComposeProviders
-}
