+++ common/pkg/config/default_unsupported.go
@@ -4,16 +4,6 @@ package config
 
 import "os"
 
-// isCgroup2UnifiedMode returns whether we are running in cgroup2 mode.
-func isCgroup2UnifiedMode() (isUnified bool, isUnifiedErr error) {
-	return false, nil
-}
-
-// getDefaultProcessLimits returns the nofile and nproc for the current process in ulimits format.
-func getDefaultProcessLimits() []string {
-	return []string{}
-}
-
 // getDefaultTmpDir for linux.
 func getDefaultTmpDir() string {
 	// first check the TMPDIR env var
