+++ common/pkg/config/default_windows.go
@@ -9,16 +9,6 @@ import (
 	"go.podman.io/storage/pkg/homedir"
 )
 
-// isCgroup2UnifiedMode returns whether we are running in cgroup2 mode.
-func isCgroup2UnifiedMode() (isUnified bool, isUnifiedErr error) {
-	return false, nil
-}
-
-// getDefaultProcessLimits returns the nofile and nproc for the current process in ulimits format
-func getDefaultProcessLimits() []string {
-	return []string{}
-}
-
 // getDefaultTmpDir for windows
 func getDefaultTmpDir() string {
 	// first check the Temp env var
